Move frame allocator to its own file
